Catch-all error handler for Kademlia transports; v0.1.9
Kad transports implement a Node EventEmitter interface, and are set up to
forward errors by emitting error events. The default behavior of Node's
EventEmitter class for `error` events is to raise them as exceptions when no
`error` event handler is registered. Asynchronous exceptions are typically not
handled, which means that the process is killed (status code 1).

This patch registers a catch-all error handler on the transport after
successful setup, so that transport errors are safely ignored.